# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ESG Carbon API",
    "description": "REST API for calculating logistics carbon footprint (CO2e) in alignment with EU CSRD / ESRS E1 disclosure requirements. Supports DEFRA 2023, 2024 and 2025 vintages; alternative fuels (HVO, SAF, LNG, methanol, electric); regional grid factors; Stripe-style idempotency; and bulk calculation.",
    "contact": {
      "name": "ESG Compliance API"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.3.0"
  },
  "paths": {
    "/api/v1/emissions/calculate": {
      "post": {
        "tags": [
          "Emissions"
        ],
        "summary": "Calculate the carbon footprint of a single shipment",
        "description": "Calculate CO2-equivalent emissions for a freight shipment.\n\nRefinements available:\n  * `sub_mode` — vehicle/vessel class (e.g. `articulated_average`).\n  * `fuel_type` — alternative fuel pathway (`hvo100`, `saf_neat`, `electric`, ...).\n  * `region` — ISO-3166 code; localises grid factor when `fuel_type='electric'`.\n  * `methodology_version` — pin to a vintage (e.g. `'2023'`) for stable historical reports.\n\nThe full derivation appears in `methodology_reference.resolution_chain`.",
        "operationId": "calculate_emissions_api_v1_emissions_calculate_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional Stripe-style idempotency key. Same key + same body → cached response (header `Idempotent-Replayed: true`). Same key + different body → 409 Conflict. TTL 24h.",
              "examples": [
                "shipment-2026-08821-attempt-1"
              ],
              "title": "Idempotency-Key"
            },
            "description": "Optional Stripe-style idempotency key. Same key + same body → cached response (header `Idempotent-Replayed: true`). Same key + different body → 409 Conflict. TTL 24h."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmissionRequest"
              },
              "examples": {
                "simple_truck": {
                  "summary": "Simple 1-tonne truck shipment, 100 km",
                  "description": "Minimal request — only the three required fields. Mode default (HGV all-average), fuel default (diesel), latest methodology.",
                  "value": {
                    "weight_kg": 1000,
                    "distance_km": 100,
                    "transport_type": "truck"
                  }
                },
                "electric_rail_france_2025": {
                  "summary": "25-tonne electric rail in France, DEFRA 2025",
                  "description": "Showcases regional grid factor: FR grid is ~12× cleaner than Poland, dragging the electric rail factor down accordingly.",
                  "value": {
                    "weight_kg": 25000,
                    "distance_km": 600,
                    "transport_type": "rail",
                    "fuel_type": "electric",
                    "region": "FR",
                    "methodology_version": "2025"
                  }
                },
                "air_freight_with_saf": {
                  "summary": "Long-haul air freight on 30% SAF blend",
                  "description": "Sustainable Aviation Fuel substitution via GLEC v3.0 §6.3 ratio method. ~25–30% emissions reduction vs neat Jet A-1.",
                  "value": {
                    "weight_kg": 500,
                    "distance_km": 9000,
                    "transport_type": "air",
                    "sub_mode": "long_haul",
                    "fuel_type": "saf_blend_30"
                  }
                },
                "container_ship_methanol": {
                  "summary": "Container ship on green methanol — alternative-fuel pathway",
                  "description": "e-methanol from renewable hydrogen (IMO MEPC.1/Circ.905). ~85–90% reduction vs heavy fuel oil baseline.",
                  "value": {
                    "weight_kg": 50000,
                    "distance_km": 8000,
                    "transport_type": "ship",
                    "sub_mode": "container",
                    "fuel_type": "methanol_green"
                  }
                },
                "pinned_historical_2023": {
                  "summary": "Pinned to DEFRA 2023 vintage (stable historical reports)",
                  "description": "When a customer's CSRD report cites a 2023-vintage figure, pin the calculation to '2023' so it never drifts as new vintages ship.",
                  "value": {
                    "weight_kg": 22000,
                    "distance_km": 850,
                    "transport_type": "truck",
                    "sub_mode": "articulated_average",
                    "methodology_version": "2023",
                    "shipment_id": "PO-2026-08821"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful calculation with full audit trail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmissionResponse"
                },
                "examples": {
                  "simple_truck": {
                    "summary": "Simple truck shipment",
                    "value": {
                      "transport_type": "truck",
                      "co2e_kg": 10.685,
                      "co2e_tonnes": 1.0685e-05,
                      "calculation": {
                        "mass_tonnes": 1.0,
                        "distance_km": 100.0,
                        "tonne_km": 100.0
                      },
                      "methodology_reference": {
                        "source": "UK DEFRA 2025 GHG Conversion Factors for Company Reporting",
                        "dataset": "Freighting goods",
                        "scope": "Well-to-Wheel (WTT + TTW combined)",
                        "unit": "kgCO2e per tonne-kilometre",
                        "version": "2025 v1.0",
                        "url": "https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2025",
                        "csrd_alignment": "ESRS E1 – Climate change. Scope 3 Cat 4 / 9.",
                        "factor_kg_co2e_per_tkm": 0.10685,
                        "sub_mode_resolved": "default",
                        "fuel_resolved": "diesel",
                        "resolution_chain": [
                          {
                            "step": "base_factor",
                            "description": "DEFRA 2025 freight factor for truck/default",
                            "value_kg_co2e_per_tkm": 0.10685,
                            "source": "UK DEFRA 2025 GHG Conversion Factors for Company Reporting"
                          }
                        ],
                        "extra_references": {
                          "version_resolved": "2025",
                          "supported_versions": [
                            "2023",
                            "2024",
                            "2025"
                          ]
                        }
                      },
                      "calculated_at": "2026-04-26T12:36:21.000Z"
                    }
                  },
                  "electric_rail_france": {
                    "summary": "Electric rail in France with regional grid adjustment",
                    "value": {
                      "transport_type": "rail",
                      "co2e_kg": 93.706452,
                      "co2e_tonnes": 9.3706e-05,
                      "calculation": {
                        "mass_tonnes": 25.0,
                        "distance_km": 600.0,
                        "tonne_km": 15000.0
                      },
                      "methodology_reference": {
                        "source": "UK DEFRA 2025 GHG Conversion Factors for Company Reporting",
                        "dataset": "Freighting goods",
                        "scope": "Well-to-Wheel (WTT + TTW combined)",
                        "unit": "kgCO2e per tonne-kilometre",
                        "version": "2025 v1.0",
                        "url": "https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2025",
                        "csrd_alignment": "ESRS E1 – Climate change. Scope 3 Cat 4 / 9.",
                        "factor_kg_co2e_per_tkm": 0.0062471,
                        "sub_mode_resolved": "electric",
                        "fuel_resolved": "electric",
                        "region_resolved": "FR",
                        "resolution_chain": [
                          {
                            "step": "base_factor",
                            "description": "DEFRA 2025 freight factor for rail/default",
                            "value_kg_co2e_per_tkm": 0.02675
                          },
                          {
                            "step": "switch_to_electric_base",
                            "description": "Replaced base factor with DEFRA rail/electric (0.02105).",
                            "value_kg_co2e_per_tkm": 0.02105
                          },
                          {
                            "step": "regional_grid_adjustment",
                            "description": "Adjusted from GB grid (0.15500) to FR grid (0.04600).",
                            "ratio": 0.29677,
                            "value_kg_co2e_per_tkm": 0.0062471
                          }
                        ],
                        "extra_references": {
                          "version_resolved": "2025",
                          "supported_versions": [
                            "2023",
                            "2024",
                            "2025"
                          ]
                        }
                      },
                      "calculated_at": "2026-04-26T12:36:21.000Z"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Idempotency-Key reused with a different request body.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Idempotency-Key 'shipment-001' was previously used with a different request body."
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed (bad input or incompatible fuel/region).",
            "content": {
              "application/json": {
                "examples": {
                  "incompatible_fuel": {
                    "summary": "Fuel/transport_type incompatibility",
                    "value": {
                      "detail": "fuel_type 'jet_a1' is not compatible with transport_type 'ship'. Compatible: ['hfo', 'lng', 'methanol_green', 'methanol_grey', 'mgo']."
                    }
                  },
                  "unknown_region": {
                    "summary": "Region not in supported list",
                    "value": {
                      "detail": "Unknown region 'ATLANTIS'. See /api/v1/reference/regions."
                    }
                  },
                  "field_validation": {
                    "summary": "Field-level validation (Pydantic)",
                    "value": {
                      "detail": [
                        {
                          "type": "greater_than",
                          "loc": [
                            "body",
                            "weight_kg"
                          ],
                          "msg": "Input should be greater than 0",
                          "input": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/emissions/modes": {
      "get": {
        "tags": [
          "Emissions"
        ],
        "summary": "List supported transport modes and sub-modes",
        "operationId": "list_modes_api_v1_emissions_modes_get",
        "responses": {
          "200": {
            "description": "Per-mode list of valid sub_mode values.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "type": "object",
                  "title": "Response List Modes Api V1 Emissions Modes Get"
                },
                "example": {
                  "truck": [
                    "articulated_average",
                    "rigid_average",
                    "van_class_iii"
                  ],
                  "rail": [
                    "diesel",
                    "electric"
                  ],
                  "ship": [
                    "bulk_carrier",
                    "container",
                    "ro_ro_ferry",
                    "tanker_general_cargo"
                  ],
                  "air": [
                    "domestic",
                    "long_haul",
                    "short_haul"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/fuels": {
      "get": {
        "tags": [
          "Reference data"
        ],
        "summary": "List fuels compatible with each transport mode",
        "operationId": "list_fuels_api_v1_reference_fuels_get",
        "responses": {
          "200": {
            "description": "Per-mode list of valid fuel_type values (compatibility matrix).",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "type": "object",
                  "title": "Response List Fuels Api V1 Reference Fuels Get"
                },
                "example": {
                  "truck": [
                    "biodiesel_b20",
                    "diesel",
                    "diesel_b7",
                    "hvo100",
                    "lng"
                  ],
                  "rail": [
                    "diesel",
                    "electric",
                    "hvo100"
                  ],
                  "ship": [
                    "hfo",
                    "lng",
                    "methanol_green",
                    "methanol_grey",
                    "mgo"
                  ],
                  "air": [
                    "jet_a1",
                    "saf_blend_30",
                    "saf_neat"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/regions": {
      "get": {
        "tags": [
          "Reference data"
        ],
        "summary": "List supported region codes and their grid factors (kgCO2e/kWh)",
        "operationId": "list_regions_api_v1_reference_regions_get",
        "responses": {
          "200": {
            "description": "Region code → grid factor (kgCO2e/kWh, EEA/EPA/IEA 2025 data).",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "number"
                  },
                  "type": "object",
                  "title": "Response List Regions Api V1 Reference Regions Get"
                },
                "example": {
                  "AT": 0.088,
                  "AU": 0.46,
                  "BE": 0.142,
                  "BR": 0.09,
                  "CA": 0.122,
                  "CH": 0.029,
                  "CN": 0.545,
                  "DE": 0.29,
                  "EU27": 0.195,
                  "FR": 0.046,
                  "GB": 0.155,
                  "PL": 0.575,
                  "US": 0.365,
                  "WORLD": 0.46
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalculationBreakdown": {
        "properties": {
          "mass_tonnes": {
            "type": "number",
            "title": "Mass Tonnes",
            "description": "weight_kg converted to tonnes."
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "tonne_km": {
            "type": "number",
            "title": "Tonne Km",
            "description": "mass_tonnes × distance_km — the activity metric."
          }
        },
        "type": "object",
        "required": [
          "mass_tonnes",
          "distance_km",
          "tonne_km"
        ],
        "title": "CalculationBreakdown",
        "description": "Intermediate values to make the result reproducible."
      },
      "EmissionRequest": {
        "properties": {
          "weight_kg": {
            "type": "number",
            "maximum": 1000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Weight Kg",
            "description": "Gross shipment weight in kilograms. Must be > 0.",
            "examples": [
              1500.0
            ]
          },
          "distance_km": {
            "type": "number",
            "maximum": 50000.0,
            "exclusiveMinimum": 0.0,
            "title": "Distance Km",
            "description": "Transport distance in kilometres. Must be > 0.",
            "examples": [
              420.5
            ]
          },
          "transport_type": {
            "$ref": "#/components/schemas/TransportType",
            "description": "Primary transport mode.",
            "examples": [
              "truck"
            ]
          },
          "sub_mode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Mode",
            "description": "Optional vehicle/vessel sub-classification (e.g. 'articulated_average', 'long_haul', 'electric'). When omitted, a sensible default for the transport_type is applied.",
            "examples": [
              "articulated_average"
            ]
          },
          "fuel_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FuelType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional fuel/energy carrier override. When provided, the base activity factor is adjusted by the fuel-substitution ratio (GLEC v3.0 §6.3) — or, for 'electric', by the regional grid factor. Must be compatible with the transport_type (e.g. 'jet_a1' is only valid for 'air').",
            "examples": [
              "hvo100"
            ]
          },
          "region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Region",
            "description": "Optional ISO-3166-1 alpha-2 country code (or 'EU27' / 'WORLD' aggregate, or 'US-CA'-style sub-national code). Used to localise the electricity grid factor when fuel_type='electric'. Ignored for non-electric fuels.",
            "examples": [
              "FR"
            ]
          },
          "methodology_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MethodologyVersion"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pin the calculation to a specific methodology vintage. Defaults to the latest published version. Pin to e.g. '2023' to guarantee stable historical results across future DEFRA updates.",
            "examples": [
              "2023"
            ]
          },
          "shipment_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipment Id",
            "description": "Optional client-supplied identifier echoed back in the response."
          }
        },
        "type": "object",
        "required": [
          "weight_kg",
          "distance_km",
          "transport_type"
        ],
        "title": "EmissionRequest",
        "description": "Input for a single shipment carbon-footprint calculation."
      },
      "EmissionResponse": {
        "properties": {
          "shipment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipment Id"
          },
          "transport_type": {
            "$ref": "#/components/schemas/TransportType"
          },
          "co2e_kg": {
            "type": "number",
            "title": "Co2E Kg",
            "description": "Total emissions in kilograms of CO2 equivalent."
          },
          "co2e_tonnes": {
            "type": "number",
            "title": "Co2E Tonnes",
            "description": "Total emissions in tonnes of CO2 equivalent."
          },
          "calculation": {
            "$ref": "#/components/schemas/CalculationBreakdown"
          },
          "methodology_reference": {
            "$ref": "#/components/schemas/MethodologyReference"
          },
          "calculated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Calculated At",
            "description": "UTC timestamp of the calculation."
          }
        },
        "type": "object",
        "required": [
          "transport_type",
          "co2e_kg",
          "co2e_tonnes",
          "calculation",
          "methodology_reference",
          "calculated_at"
        ],
        "title": "EmissionResponse",
        "description": "Carbon-footprint calculation result for a single shipment."
      },
      "FuelType": {
        "type": "string",
        "enum": [
          "diesel",
          "diesel_b7",
          "hvo100",
          "biodiesel_b20",
          "lng",
          "hfo",
          "mgo",
          "jet_a1",
          "saf_blend_30",
          "saf_neat",
          "methanol_grey",
          "methanol_green",
          "electric"
        ],
        "title": "FuelType",
        "description": "Operator-specified fuel/energy carrier — overrides the mode default."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MethodologyReference": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "dataset": {
            "type": "string",
            "title": "Dataset"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "csrd_alignment": {
            "type": "string",
            "title": "Csrd Alignment"
          },
          "factor_kg_co2e_per_tkm": {
            "type": "number",
            "title": "Factor Kg Co2E Per Tkm",
            "description": "Final emission factor applied (after any fuel/region adjustments)."
          },
          "sub_mode_resolved": {
            "type": "string",
            "title": "Sub Mode Resolved",
            "description": "The sub_mode that was actually applied (after defaulting)."
          },
          "fuel_resolved": {
            "type": "string",
            "title": "Fuel Resolved",
            "description": "The fuel/energy carrier the calculation effectively used."
          },
          "region_resolved": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Resolved",
            "description": "The region used for grid-factor lookup, if any."
          },
          "resolution_chain": {
            "items": {
              "$ref": "#/components/schemas/ResolutionStep"
            },
            "type": "array",
            "title": "Resolution Chain",
            "description": "Step-by-step derivation of the final factor — base DEFRA factor → fuel substitution → regional grid adjustment. Reproducible by hand."
          },
          "extra_references": {
            "type": "object",
            "title": "Extra References",
            "description": "Supplementary methodology fields (e.g. fuel-substitution method, grid source)."
          }
        },
        "type": "object",
        "required": [
          "source",
          "dataset",
          "scope",
          "unit",
          "version",
          "url",
          "csrd_alignment",
          "factor_kg_co2e_per_tkm",
          "sub_mode_resolved",
          "fuel_resolved"
        ],
        "title": "MethodologyReference",
        "description": "Auditable provenance of the emission factor used."
      },
      "MethodologyVersion": {
        "type": "string",
        "enum": [
          "latest",
          "2023",
          "2024",
          "2025"
        ],
        "title": "MethodologyVersion",
        "description": "Pin the calculation to a specific methodology vintage. ``LATEST`` always\ntracks the most recent published vintage; an explicit year is frozen so\nhistorical CSRD reports stay reproducible across DEFRA updates."
      },
      "ResolutionStep": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step",
            "description": "Machine-readable step identifier."
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "value_kg_co2e_per_tkm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Kg Co2E Per Tkm"
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "step",
          "description"
        ],
        "title": "ResolutionStep",
        "description": "One step in the factor-derivation chain — preserved for auditors."
      },
      "TransportType": {
        "type": "string",
        "enum": [
          "truck",
          "ship",
          "air",
          "rail"
        ],
        "title": "TransportType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}